Instrument AST to help with JIT
Implements four modifications to AST: adds global variable to return
result, makes last expression be assignment to this global variable,
wraps user's expression into a function, makes last expression be a
return statement, and makes all declarations public. This is so that
we can lookup this function in JIT and call it, and simply print out
whatever that function returns. This change also updates linkages in
the SIL module for our generated function and for main (to prevent
main from being generated in the IR).